Refactored TextureWrapper code
This was motivated by making it easier to add new texture parameters which are pass through to the GPU.
This eliminates some duplicated code at the cost of always having to call
 TextureWrapper.loadAndInstallAllTexture() before running  chronosGL.run();
Before you could get away without doing this if you only used Canvas texture.

Also got rid of the TextureCache - these are now folded into the TextureWrapper as static functions.

Most of the FX examples are still badly broken - except for plasma